Move persistence files in data folder
