+++ app/src/main/java/com/vaudibert/canidrive/data/DrinkEntity.kt
@@ -1,4 +1,4 @@
-package com.vaudibert.canidrive
+package com.vaudibert.canidrive.data
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
